requirements.txt
tqdm
atlassian-python-api
python-dotenv
iso8601
